requirements.txt
langchain
unstructured
openai
Cython
tiktoken
langchain-astradb
langchain_community
pdf2image
pdfminer.six
unstructured[pdf]
pypdf
langchain-openai
langchain-mongodb
datasets
numpy 
pandas
pymongo
langchain_community
pypdf
faiss-cpu
cryptography>=3.1
python-dotenv
pinecone-client
langchain_pinecone
flask
openpyxl
Dataset
ragas